% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.adj_cif.R
\name{adj_cif}
\alias{adj_cif}
\title{use R code to calculate adjusted CIF}
\usage{
adj_cif(
  data,
  time,
  status,
  group,
  covlist,
  event_code,
  stratified = "Yes",
  reference_group = NULL
)
}
\arguments{
\item{data}{input data}

\item{time}{time}

\item{status}{status}

\item{group}{grouping variable}

\item{covlist}{adjusted covariates}

\item{event_code}{event of interests}

\item{stratified}{"Yes" refers to use stratified model, "No" refers to use FG}

\item{reference_group}{NULL- unstratified FG when stratified = No; "G&B"- G&B when stratified = Yes;}
}
\value{
a dataframe
}
\description{
Use input data, time, status,grouping variables, adjusted covariates,
events of interests, whether to use stratified model, and defining reference group as inputs
}
\keyword{internal}
