% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.adjust_CIF_plot.R
\name{adjCIF_CI_plot}
\alias{adjCIF_CI_plot}
\title{Generate adjusted CIF plot with bootstrap CI}
\usage{
adjCIF_CI_plot(res, data)
}
\arguments{
\item{res}{results from boot_ci_adj_cif()}

\item{data}{the input dataset}
}
\value{
Adjusted CIF plot with CI will be shown after running this function
}
\description{
Use results from boot_ci_adj_cif() and input dataset to generate adjusted CIF plot with bootstrap CI (users can also produce a figure by using ggplot2)
}
\examples{

library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm = factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 = as.character(bmt$z3)

bmt$CenCI <- 0
for (ii in 1:137) {
  if (bmt$d3[ii] == 0) {
    bmt$CenCI[ii] <- 0
  } else {
    if (bmt$d2[ii] == 1) {
      bmt$CenCI[ii] <- 1
    } else {
      bmt$CenCI[ii] <- 2
    }
  }
}

bmt$t2 = bmt$t2 * 12/365.25

# Adjusted CIF plot with bootstrap CI
result = boot_ci_adj_cif(boot_n = 100, ci_cut = c(0.025, 0.975), data = bmt, time = "t2",
status = "CenCI", group = "arm", covlist = c("z1", "z3"), event_code = 1, "No",
                            NULL)
adjCIF_CI_plot(result, data = bmt)


}
