% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.adjust_CIF_plot.R
\name{adjCIF_plot}
\alias{adjCIF_plot}
\title{Generate adjusted CIF plot}
\usage{
adjCIF_plot(res, data)
}
\arguments{
\item{res}{results from adjusted_CIF()}

\item{data}{the input dataset}
}
\value{
a figure
}
\description{
Use results from adjusted_CIF() and input dataset as inputs
}
\examples{
result = adjusted_CIF(data = bmt, time = "t2", status = "CenCI", group = "arm",
covlist = c("z1", "z3"), event_code = 1, stratified = "Yes", reference_group = "arm:2")
adjCIF_plot(result, data = bmt)
}
