% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_adjust_KM_plot.R
\name{adjKM_CI_plot}
\alias{adjKM_CI_plot}
\title{Generate adjusted KM plot with bootstrap CI}
\usage{
adjKM_CI_plot(res, data)
}
\arguments{
\item{res}{results from boot_ci_adj_km()}

\item{data}{the input dataset}
}
\value{
Adjusted CIF plot with CI will be shown after running this function
}
\description{
Use results from boot_ci_adj_km() and input dataset to generate adjusted KM plot with bootstrap CI (users can also produce a figure by using ggplot2)
}
\examples{
library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm = factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 = as.character(bmt$z3)
bmt$t2 = bmt$t2 * 12/365.25

result1_1 = boot_ci_adj_km(boot_n = 100, ci_cut = c(0.025, 0.975), data = bmt, time = "t2",
status = "d3", group = "arm", covlist = c("z1", "z3"), stratified_cox = "No",
reference_group = NULL)

adjKM_CI_plot(result1_1, bmt)
}
