% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_adjust_KM_plot.R
\name{adjKM_plot}
\alias{adjKM_plot}
\title{Generate adjusted KM plot without CI}
\usage{
adjKM_plot(res, data)
}
\arguments{
\item{res}{results from adjusted_KM()}

\item{data}{the input dataset}
}
\value{
Adjusted KM plot will be shown after running this function
}
\description{
Use results from adjusted_KM() and input dataset to generate adjusted survival plot (users can also produce a figure by using ggplot2)
}
\examples{
# Data preparation
library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm = factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 = as.character(bmt$z3)
bmt$t2 = bmt$t2 * 12/365.25

# Cox model
result1 = adjusted_KM(data = bmt, time = "t2", status = "d3", group = "arm", covlist = c("z1",
"z3"), stratified_cox = "No", reference_group = NULL)
adjKM_plot(result1, data = bmt)

}
