% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_CIF_func.R
\name{adjusted_CIF}
\alias{adjusted_CIF}
\title{Covariate-Adjusted cumulative incidence functions}
\usage{
adjusted_CIF(
  data,
  time,
  status,
  group,
  covlist,
  event_code = 1,
  stratified = "Yes",
  reference_group = "G&B"
)
}
\arguments{
\item{data}{the input data set}

\item{time}{column name of time variable}

\item{status}{column name of event status}

\item{group}{grouping variable}

\item{covlist}{list of covariates that should be included in the model}

\item{event_code}{event of interests that indicates the failure type of interest}

\item{stratified}{"Yes" refers to use stratified model, "No" refers to use Fine and Gray regression}

\item{reference_group}{“NULL” - No reference group required for the Cox PH model;“G&B” - the Gail and Byar method; “group:level” - the reference group for the Storer method (e.g., “arm:2” in the BMT data)}
}
\value{
The function generates a data frame that contains the adjusted CIF probabilities. In cases where the proportional hazards (PH) assumption is not valid, or if practitioners require the event time points of the adjusted function to match those of the unadjusted function, the stratified model (using Gail and Byar or Storer et al. methods) should be used. Otherwise, the unstratified Fine-Gray (FG) model is sufficient. Additional information about this can be found in the "adjusted_KM()" function.
}
\description{
The function computes the covariate-adjusted CIF based on Fine-Gray or stratified Fine-Gray model. Three approaches are implemented. See the detail in "adjusted_KM()" function.
}
\examples{

\dontrun{
library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm <- factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 <- as.character(bmt$z3)

bmt$CenCI <- 0
for (ii in 1:137) {
  if (bmt$d3[ii] == 0) {
    bmt$CenCI[ii] <- 0
  } else {
    if (bmt$d2[ii] == 1) {
      bmt$CenCI[ii] <- 1
    } else {
      bmt$CenCI[ii] <- 2
    }
  }
}

bmt$t2 <- bmt$t2 * 12 / 365.25

# unstratified FG
result1 <- adjusted_CIF(
  data = bmt, time = "t2", status = "CenCI", group = "arm",
  covlist = c("z1", "z3"), event_code = 1, stratified = "No", reference_group = NULL
)
# stratified G&B
result2 <- adjusted_CIF(
  data = bmt, time = "t2", status = "CenCI", group = "arm",
  covlist = c("z1", "z3"), event_code = 1, stratified = "Yes", reference_group = "G&B"
)
# stratified Storer
result3 <- adjusted_CIF(
  data = bmt, time = "t2", status = "CenCI", group = "arm",
  covlist = c("z1", "z3"), event_code = 1, stratified = "Yes", reference_group = "arm:2"
)
}
}
