% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_KM_func.R
\name{adjusted_KM}
\alias{adjusted_KM}
\title{Covariate-adjusted survival probability based on Cox/stratified Cox model}
\usage{
adjusted_KM(
  data,
  time,
  status,
  group,
  covlist,
  stratified_cox = "Yes",
  reference_group = "G&B"
)
}
\arguments{
\item{data}{the input dataset}

\item{time}{column name of time variable}

\item{status}{column names of event status, 0 is the censored and 1 is the event}

\item{group}{group variable being compared}

\item{covlist}{list of covariates included in the Cox model}

\item{stratified_cox}{"Yes" refers to use stratified model, "No" refers to use Coxph}

\item{reference_group}{"NULL" - No reference group required for the Cox PH model;"G&B" - the Gail and Byar method; "group:level" - the reference group for the Storer method (e.g., "arm:2" in the BMT data)}
}
\value{
The output of this function is a dataframe that contains the adjusted survival probabilities. If you specify "stratified_cox="Yes"", the function will apply a stratified Cox regression model, where the group variable serves as the stratification variable. You can choose the reference group using the Gail and Byar method by specifying "reference_group = "G&B"", which includes all subjects in each group, or using the Storer method by specifying "reference_group = "variable_name:level"", which selects a specific group as the reference group.
If you set "stratified_cox = "No"" and "reference_group = NULL", the function will use the unstratified Cox model to estimate the adjusted survival probability. This is appropriate if the PH assumption is valid, as it uses a common baseline survival function and there is no difference in event time points between groups. However, the number of events may not match the actual number of events in each group.
If you need to ensure that the event time points of the adjusted function match those of the unadjusted function, you can choose the stratified Cox model as an alternative.
}
\description{
This function calculates the adjusted Kaplan-Meier functions by taking into account the covariates, using either Cox or stratified Cox model. You can find further information about this in the publication titled "AdjKM.CIF: An R package for estimating the covariate-adjusted Kaplan-Meier and cumulative incidence functions". The function offers three different approaches, which are discussed in the value section.
}
\examples{
# Data preparation
\dontrun{
library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm <- factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 <- as.character(bmt$z3)
bmt$t2 <- bmt$t2 * 12 / 365.25

# Cox model
result1 <- adjusted_KM(data = bmt, time = "t2", status = "d3", group = "arm", covlist = c(
  "z1",
  "z3"
), stratified_cox = "No", reference_group = NULL)

# Stratified cox: Gail&Byar's approach
result2 <- adjusted_KM(data = bmt, time = "t2", status = "d3", group = "arm", covlist = c(
  "z1",
  "z3"
), stratified_cox = "Yes", reference_group = "G&B")

# Stratified cox: Storer's approach
result3 <- adjusted_KM(data = bmt, time = "t2", status = "d3", group = "arm", covlist = c(
  "z1",
  "z3"
), stratified_cox = "Yes", reference_group = "arm:2")
}

}
