% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.baseline_function_CIF.R
\name{baseline_hazard_cif}
\alias{baseline_hazard_cif}
\title{use R code to calculate baseline CIF}
\usage{
baseline_hazard_cif(data, time, status, group, covlist, event_code, beta)
}
\arguments{
\item{data}{input data}

\item{time}{time}

\item{status}{status}

\item{group}{grouping variable}

\item{covlist}{adjusted covariates}

\item{event_code}{event of interests}

\item{beta}{beta from FG or stratified model}
}
\value{
a dataframe
}
\description{
Use input data, time, status,grouping variables, adjusted covariates,
events of interests, the known beta as inputs
}
\keyword{internal}
