% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCI_KM.R
\name{boot_ci_adj_km}
\alias{boot_ci_adj_km}
\title{Bootstrap CI for adjusted KM}
\usage{
boot_ci_adj_km(
  boot_n = 100,
  ci_cut = c(0.025, 0.975),
  data,
  time,
  status,
  group,
  covlist,
  stratified_cox,
  reference_group
)
}
\arguments{
\item{boot_n}{bootstrap sample size}

\item{ci_cut}{default c(0.025, 0.975) bootstrap 95\% CI}

\item{data}{the input dataset}

\item{time}{column name of time variable}

\item{status}{column name of event status}

\item{group}{grouping variable}

\item{covlist}{list of covariates that should be included in the model}

\item{stratified_cox}{"Yes" refers to use stratified model, "No" refers to use coxph regression}

\item{reference_group}{"NULL" - No reference group required for the Cox PH model;"G&B" - the Gail and Byar method; "group:level" - the reference group for the Storer method (e.g., "arm:2" in the BMT data)}
}
\value{
Output is a dataframe with average number of adjusted survival probabilities, as well as 2.5\% and 97.5\% percentiles.
}
\description{
Computes the covariate-adjusted KM functions along with confidence intervals by the bootstrap percentile method.
}
\examples{
# Data preparation
\dontrun{
library(KMsurv)
data(bmt)
bmt$arm <- bmt$group
bmt$arm <- factor(as.character(bmt$arm), levels = c("2", "1", "3"))
bmt$z3 <- as.character(bmt$z3)
bmt$t2 <- bmt$t2 * 12 / 365.25

# Unstratified cox
result1_1 <- boot_ci_adj_km(
  boot_n = 100, ci_cut = c(0.025, 0.975), data = bmt, time = "t2",
  status = "d3", group = "arm", covlist = c("z1", "z3"), stratified_cox = "No",
  reference_group = NULL
)

# Stratified Cox: Gail and Byar's method
result1_2 <- boot_ci_adj_km(
  boot_n = 100, ci_cut = c(0.025, 0.975), data = bmt, time = "t2",
  status = "d3", group = "arm", covlist = c("z1", "z3"), stratified_cox = "Yes",
  reference_group = "G&B"
)

# Stratified Cox: Storer's approach
result1_3 <- boot_ci_adj_km(
  boot_n = 100, ci_cut = c(0.025, 0.975), data = bmt, time = "t2",
  status = "d3", group = "arm", covlist = c("z1", "z3"), stratified_cox = "Yes",
  reference_group = "arm:2"
)
}
}
