% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. cif_prob.R
\name{cif_prob}
\alias{cif_prob}
\title{use R code to calculate adjusted CIF probability (internal function)}
\usage{
cif_prob(
  data,
  coxout,
  base_res,
  event_code,
  strata_group,
  stratified,
  reference_group
)
}
\value{
a dataframe
}
\description{
Use input data, time, status,grouping variables, adjusted covariates,
events of interests, whether to use stratified model, and defining reference group as inputs
}
\keyword{internal}
