% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.adjusted_CIF_final.R
\name{adjusted_CIF}
\alias{adjusted_CIF}
\title{use R code to calculate the adjusted CIF}
\usage{
adjusted_CIF(
  data,
  time,
  status,
  group,
  covlist,
  event_code = 1,
  stratified = "Yes",
  reference_group = "G&B"
)
}
\arguments{
\item{data}{the input dataset}

\item{time}{column name of time variable}

\item{status}{column name of event status}

\item{group}{grouping variable}

\item{covlist}{list of covariates that are included in the model}

\item{event_code}{event of interests}

\item{stratified}{"Yes" refers to use stratified model, "No" refers to use FG}

\item{reference_group}{NULL- unstratified FG when stratified = No; "G&B"- G&B when stratified = Yes;}
}
\value{
a dataframe
}
\description{
Use input data, time, status,grouping variables, adjusted covariates,
events of interests, whether to use stratified model, and defining reference group as inputs
}
\examples{

# unstratified FG
result1 = adjusted_CIF(data = bmt, time = "t2", status = "CenCI", group = "arm",
covlist = c("z1", "z3"), event_code = 1, stratified = "No", reference_group = NULL)
# stratified G$B
result2 = adjusted_CIF(data = bmt, time = "t2", status = "CenCI", group = "arm",
covlist = c("z1", "z3"), event_code = 1, stratified = "Yes", reference_group = "G&B")
# stratified Storer
result3 = adjusted_CIF(data = bmt, time = "t2", status = "CenCI", group = "arm",
covlist = c("z1", "z3"), event_code = 1, stratified = "Yes", reference_group = "arm:2")
}
